refactor(framework): code quality review fixes
- remove initialize dead code: AbstractApplicationCommandLineRunnerHandler (0 subclasses),
  ApplicationCommandLineRunnerExecute, InitializeHandlerType dead constants
- remove redundant @WebFilter on BackManageAuthFilter (filter is registered by FilterAutoConfiguration @Bean)
- remove ineffective @Order on FilterAutoConfiguration @Bean methods (does not affect ordering; actual order = declaration order)
- fix BasePageDto @Schema type Long->Integer
- ServerGrayFilter: add getMetadata null guard (consistent with L69)
- GatewayContextAutoConfiguration bean method webMvcContext->gatewayContextHandler
- fix fiterbalance package typo to filterbalance

+++ ticketflow-spring-cloud-framework/ticketflow-service-common/src/main/java/com/ticketflow/dto/BasePageDto.java
@@ -13,12 +13,12 @@ import lombok.Data;
 public class BasePageDto {
 
 
-    @Schema(name = "pageNumber", type = "Long", description = "页码", requiredMode = RequiredMode.REQUIRED)
+    @Schema(name = "pageNumber", type = "Integer", description = "页码", requiredMode = RequiredMode.REQUIRED)
     @NotNull
     private Integer pageNumber;
 
 
-    @Schema(name = "pageSize", type = "Long", description = "页大小", requiredMode = RequiredMode.REQUIRED)
+    @Schema(name = "pageSize", type = "Integer", description = "页大小", requiredMode = RequiredMode.REQUIRED)
     @NotNull
     private Integer pageSize;
 }
